break up files into types.rs and sequential.rs
